+++ messaging_api/src/domain/message/message.go
@@ -9,8 +9,8 @@ type Message struct {
 	AuthorId         int64   `json:"author_id"` //Not needed anymore since messages will be searched from conversations
 	AuthorUuid       string  `json:"author_uuid"`
 	Text             string  `json:"text"`
-	CreatedAt        float32 `json:"created_at"`
-	UpdatedAt        float32 `json:"updated_at"`
+	CreatedAt        float64 `json:"created_at"`
+	UpdatedAt        float64 `json:"updated_at"`
 }
 
 func (message *Message) SetAuthorId(id int64) {
